+++ api/v1alpha/src/accounts/dataManager.js
@@ -507,21 +507,6 @@ async function reset(projectId, accountId) {
     }
 }
 
-/**
- * @param  {} projectId
- * @param  {} entitlementId
- */
-async function autoApproveEntitlement(projectId, entitlementId) {
-    const procurementUtil = new CommerceProcurementUtil(projectId);
-    const entitlementName = procurementUtil.getEntitlementName(projectId, entitlementId);
-    const entitlement = await procurementUtil.getEntitlement(entitlementName);
-    console.log(entitlement);
-    const product = entitlement.product;
-    const plan = entitlement.plan;
-    console.log(`Auto approve product: ${product} and plan: ${plan}`);
-    await procurementUtil.approveEntitlement(entitlementName);
-}
-
 module.exports = {
     listAccounts,
     createOrUpdateAccount,
@@ -529,6 +514,5 @@ module.exports = {
     getAccount,
     register,
     activate,
-    reset,
-    autoApproveEntitlement
+    reset
 };
